+++ src/Filter/AbstractFilter.php
@@ -85,10 +85,5 @@ abstract class AbstractFilter implements FilterInterface
     /**
      * @inheritDoc
      */
-    abstract public function filterResults(array $facetedData, ?array $inputIdKeys = null): array;
-
-    /**
-     * @inheritDoc
-     */
     abstract public function filterInput(array $facetedData,  array &$inputIdKeys): void;
 }