+++ src/Filter/FilterInterface.php
@@ -37,13 +37,7 @@ interface FilterInterface
      * @return string
      */
     public function getFieldName(): string;
-    /**
-     * Filter faceted data
-     * @param array<int|string,array<int>|\SplFixedArray<int>> $facetedData
-     * @param array<int,bool|int>|null $inputIdKeys - RecordId passed into keys of an array (performance issue)
-     * @return array<int,bool> - results in keys
-     */
-    public function filterResults(array $facetedData, ?array $inputIdKeys = null): array;
+    
     /**
      * Filter faceted data
      * @param array<int|string,array<int>|\SplFixedArray<int>> $facetedData